ouel_django/requirements/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.3
distro==1.9.0
Django==6.0.1
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-oauth-toolkit==3.2.0
djangorestframework==3.16.1
djecrety==1.0.17
drf-yasg==1.21.14
google-auth==2.47.0
google-genai==1.59.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.3.1
packaging==25.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
